Moved EventChannel creation in the downloadOfflineRegion method (#205)
Here i splitted the downloadOfflineRegion method on the native side both
on IOS and android. This is to avoid edge case when the download is so
fast that the event channel doesnt link to the native side. When this
happen, the callback `onEvent` is never called. This is an inconsitency
that can create infinite loadings.