Gps example page (#578)
Hey!
I've added a GPS example page with 2 buttons.
The first button enabled GPS and navigates to the current position, and
the second button switches from the default LocationEnginePlatforms
settings to a Setting with LocationPriority.highAccuracy (on Android)

The location behavior is different on android. 
Whenever you keep the default setting, maplibre does not update the
location. (left you can see an active navigation - right the current
position in Maplibre - it's clearly off)
![Screenshot 2025-05-12 at 16 12
32](https://github.com/user-attachments/assets/96a4907e-f021-4132-a4b1-868b7e341b71)

whenever i activate high accuracy, the position actually updates and is
following the user.
![Screenshot 2025-05-12 at 16 12
44](https://github.com/user-attachments/assets/f1f0b40f-2730-48a6-acfc-e3e3374ff693)

---------

Co-authored-by: alexanderthiele <alexander.thiele@ocell.io>