Fixed ios crashes related to source and layers changes (#574)
This PR addresses several crashes that occurred when attempting to add
layers and sources that already exist.


The fix introduces appropriate guard statements and ensures that errors
are correctly propagated to the Flutter layer when necessary.

Previously, such issues would result in a crash. With this change, they
now raise a platform exception instead.

---------

Co-authored-by: Alexander Thiele <AlexanderThiele@users.noreply.github.com>