Prevented calling notifyListeners() after the controller is disposed. (#621)
#### Issue:

On mobile, if an async action such as adding or removing features is
executed right before the controller is disposed, an error may occur.
This happens because these actions are asynchronous, and the controller
still attempts to call notifyListeners() after it has already been
disposed.

#### Fix:

• Introduced an isDisposed flag inside the controller.
• Wrapped calls to notifyListeners() with a check to ensure the
controller has not been disposed.